refactor: separate path transform test
